backend/requirements.txt
# NeuroInspect Backend Dependencies
# Core Framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.5
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Computer Vision & Deep Learning
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
scikit-image>=0.21.0
Pillow>=10.0.0

# Explainability
grad-cam>=1.4.0

# Clustering & Analytics
scikit-learn>=1.3.0
hdbscan>=0.8.30

# Database & Storage
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
httpx>=0.25.0
aiofiles>=23.0.0

# Image Processing
imageio>=2.30.0
